+++ services/api/src/lib.rs
@@ -8,14 +8,18 @@
 use rocket::{Build, Rocket};
 use rocket_cors::CorsOptions;
 use routes::register::register_user;
+use storage_service::init_db;
 
 #[macro_use]
 extern crate rocket;
 
 #[launch]
 pub async fn rocket() -> Rocket<Build> {
+	// TODO: don't hardcode here.
+	let connection = init_db("users.db").expect("Failed to init db connection");
 	rocket::build()
 		.attach(CorsOptions::default().to_cors().unwrap())
+		.manage(connection)
 		.mount("/", routes![register_user])
 }
